自动升级

+++ AU.Common/AppRemotePublish.cs
@@ -85,15 +85,24 @@ namespace AU.Common
                 return 1;
             if (remote == null)
                 return -1;
+            string filepath = this.LocalPath + "\\" + subsystem + "\\" + local.DownPath;
             //有更新
             if (remote.No.CompareTo(local.No) > 0)
             {
                 return 1;
             }
+            else if (!System.IO.File.Exists(filepath))
+            //判断本地文件是否存在，且hash值相同
+            {
+                return 1;
+            }
+            else if (ToolsHelp.ComputeSHA256(filepath).ToLower() != remote.SHA256.ToLower())
+            {
+                return 1;
+            }
 
             return 0;
-        }
-
+        }      
         /// <summary>
         /// 通知消息
         /// </summary>
@@ -110,8 +119,9 @@ namespace AU.Common
         /// </summary>
         /// <param name="upgradeFiles"></param>
         /// <returns></returns>
-        public string DownUpdateFile(string subSystem, AuPublish upgradeFiles, bool allowPublish = false)
+        public string DownUpdateFile(string subSystem, AuPublish upgradeFiles, out AuPublish notify, bool allowPublish = false)
         {
+            notify = null;
             if (upgradeFiles == null)
             {
                 return string.Empty;
@@ -122,36 +132,40 @@ namespace AU.Common
                 int index = 0;
                 string down = string.Format("{0}{1}/{2}", upgradeFiles.Url.StartsWith("http://", StringComparison.InvariantCultureIgnoreCase) ? "" : "http://", upgradeFiles.Url, upgradeFiles.DownPath);
                 NotifyMessage(new Common.NotifyMessage(NotifyType.StartDown, "开始下载文件"));
-
-                string tempPath = Environment.GetEnvironmentVariable("Temp") + "\\" + "_" + upgradeFiles.SHA256 + "\\" + upgradeFiles.DownPath;
+                string temproot = Environment.GetEnvironmentVariable("Temp") + "\\AuUpdate\\" + subSystem;
+                string tempPath = temproot + "\\" + upgradeFiles.SHA256 + "\\" + upgradeFiles.DownPath;
+                //删除临时目录
+                ToolsHelp.DeleteDirectory(temproot, upgradeFiles.SHA256);
                 ToolsHelp.CreateDirtory(tempPath);
-
-                WebRequest webReq = WebRequest.Create(down);
-                WebResponse webRes = webReq.GetResponse();
-
-                NotifyMessage(new Common.NotifyMessage(NotifyType.Process, "正在下载[" + upgradeFiles.DownPath + "]文件,请稍后...", 100000));
-                Stream srm = webRes.GetResponseStream();
-                Stream outStream = System.IO.File.Create(tempPath);
-                byte[] buffer = new byte[1024];
-                int l;
-                int startByte = 0;
-                do
+                if (!(System.IO.File.Exists(tempPath) && ToolsHelp.ComputeSHA256(tempPath).ToLower() == upgradeFiles.SHA256.ToLower()))
                 {
-                    l = srm.Read(buffer, 0, buffer.Length);
-                    if (l > 0)
+                    WebRequest webReq = WebRequest.Create(down);
+                    WebResponse webRes = webReq.GetResponse();
+
+                    NotifyMessage(new Common.NotifyMessage(NotifyType.Process, "正在下载[" + upgradeFiles.DownPath + "]文件,请稍后...", 100000));
+                    Stream srm = webRes.GetResponseStream();
+                    Stream outStream = System.IO.File.Create(tempPath);
+                    byte[] buffer = new byte[1024];
+                    int l;
+                    int startByte = 0;
+                    do
                     {
-                        startByte += l;
-                        outStream.Write(buffer, 0, l);
-                        NotifyMessage(new Common.NotifyMessage(NotifyType.UpProcess, index + ":" + Convert.ToInt32((startByte / 100000) * 100).ToString() + "%", l));
+                        l = srm.Read(buffer, 0, buffer.Length);
+                        if (l > 0)
+                        {
+                            startByte += l;
+                            outStream.Write(buffer, 0, l);
+                            NotifyMessage(new Common.NotifyMessage(NotifyType.UpProcess, index + ":" + Convert.ToInt32((startByte / 100000) * 100).ToString() + "%", l));
+                        }
                     }
+                    while (l > 0);
+                    outStream.Flush();
+                    outStream.Close();
+                    srm.Close();
+                    index++;
                 }
-                while (l > 0);
-                outStream.Flush();
-                outStream.Close();
-                srm.Close();
-                index++;
                 if (allowPublish)
-                    return this.Publis(subSystem, upgradeFiles, tempPath);
+                    return this.Publis(subSystem, upgradeFiles, tempPath, out notify);
 
                 return tempPath;
             }
@@ -171,19 +185,22 @@ namespace AU.Common
         /// <param name="auPublish"></param>
         /// <param name="filePath"></param>
         /// <returns></returns>
-        public string Publis(string subSystem, AuPublish auPublish, string filePath)
+        public string Publis(string subSystem, AuPublish auPublish, string filePath, out AuPublish notify)
         {
             //发布地址转换]
             auPublish.Url = this.PublishAddress + "/" + subSystem;
-            string targetPath = this.LocalPath + "\\package\\" + subSystem + "\\" + auPublish.DownPath;
+            string targetPath = this.LocalPath + subSystem + "\\" + auPublish.DownPath;
             ToolsHelp.CreateDirtory(targetPath);
             File.Copy(filePath, targetPath, true);
             //写发布
-            StreamWriter swau = new StreamWriter(this.LocalPath + "\\package\\" + subSystem + "\\" + AppPublish.PackageName, false, System.Text.Encoding.UTF8);
+            StreamWriter swau = new StreamWriter(this.LocalPath + subSystem + "\\" + AppPublish.PackageName, false, System.Text.Encoding.UTF8);
             swau.Write(Newtonsoft.Json.JsonConvert.SerializeObject(auPublish));
             swau.Close();
             //删除临时目录
-            System.IO.Directory.Delete(System.IO.Path.GetDirectoryName(filePath), true);
+            //System.IO.Directory.Delete(System.IO.Path.GetDirectoryName(filePath), true);
+            //删除不必要文件
+            ToolsHelp.DeleteFile(System.IO.Path.GetDirectoryName(targetPath), true, auPublish.DownPath, AppPublish.PackageName);
+            notify = auPublish;
 
             return targetPath;
         }
